version location, sha256, comments
new location to get current version from
checking sha256 of downloaded binary
more comments
less output

+++ upgrade_gitea.sh
@@ -1,40 +1,60 @@
 #!/bin/bash
+BASEURL=https://dl.gitea.io/gitea
 ARCH="linux-amd64"
 GITEA_BIN_DIR="/usr/local/bin"
 
+#cd into directory where the script lives
 cd "$( cd "$( dirname "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )"
 
+#Get version number of latest release
 if [ -z "$1" ]
   then
-    VERSION=$(curl -L -s -S https://dl.gitea.io/charts/index.yaml | grep appVersion | head -n 1 | awk '{print $2}')
+    VERSION=$(curl -L -s -S $BASEURL -o - | grep '>Current Release' | awk '{split($0,a,"Release "); split(a[2],b,"<")} END{print b[1]}')
   else
     VERSION=$1
 fi
+
+#Check version is available by test-downloading the sha256
+URL=$BASEURL/$VERSION/gitea-$VERSION-$ARCH
+CODE=$(curl -L -s -S -w "%{http_code}" -O $URL.sha256)
+
+if [ "$CODE" -ne 200 ]
+	then
+		echo "Version $VERSION does not exist online!"
+		exit 1
+fi
+
+
 # CURRENT=`ls gitea*$ARCH | tail -n 1 | awk '{split($0,array,"-")} END{print array[2]}'` #if older files are in same folder and api is disabled
 CURRENT=$(curl -s -X GET "http://localhost:3000/api/v1/version" -H  "accept: application/json" | jq -r '.version')
 
 echo -e "Your current gitea version seems to be: $CURRENT\nGitea version available online: $VERSION"
 if dpkg --compare-versions "$VERSION" gt "$CURRENT"
 	then
-		echo -e "Your version seems to be outdated.\nDo you want to upgrade? (Need to enter password for sudo)"
+		echo -e "Your version seems to be outdated.\nDo you want to upgrade? (Need to enter password)"
 		read -rn 1 YN
 		echo ""
 		if [ "$YN" == "y" ] || [ "$YN" == "Y" ] || [ "$YN" == "z" ] || [ "$YN" == "Z" ]
 			then
-				URL=https://dl.gitea.io/gitea/$VERSION/gitea-$VERSION-$ARCH
 				FILENAME="$(basename "$URL")"
 				echo "Downloading file ..."
 				curl -L -O "$URL"
-				echo "This is what we've downloaded:"
-				ls -lLah ./"$FILENAME"
+				echo "Comparing sha256s. This may take a while ..."
+				SUMSUM=$(cat $FILENAME.sha256)
+				FILESUM=$(sha256sum $FILENAME)
+				if [ "$SUMSUM" = "$FILESUM" ]
+					then
+						echo "sha256s are identical"
+					else
+						echo "sha256s differ! Aborting."
+						exit 1
+				fi
 				echo "Backing up gitea."
 				sudo cp $GITEA_BIN_DIR/gitea ./gitea
 				echo "Stopping gitea..."
 				sudo systemctl stop gitea
 				echo "Replacing executable."
 				sudo cp "$FILENAME" $GITEA_BIN_DIR/gitea
-				echo "This is what we've installed:"
-				sudo ls -lLah $GITEA_BIN_DIR/gitea
 				echo "Starting gitea again."
 				sudo systemctl start gitea
 			else
